+++ common/core/src/test/java/is/codion/common/resource/TestResources.java
@@ -18,17 +18,10 @@
  */
 package is.codion.common.resource;
 
-/**
- * Provides overidden resource messages.
- */
-public interface Messages {
+public final class TestResources implements Resources {
 
-	/**
-	 * Returns a value for overriding the default resource value or the default string if no override is provided
-	 * @param baseBundleName the base resource bundle name
-	 * @param key the key
-	 * @param defaultString the default string
-	 * @return a value for overriding the default value or the default string if no override is provided
-	 */
-	String get(String baseBundleName, String key, String defaultString);
+	@Override
+	public String get(String baseBundleName, String key, String defaultString) {
+		return defaultString + " override";
+	}
 }