+++ demos/chinook/src/main/java/module-info.java
@@ -16,6 +16,7 @@
  *
  * Copyright (c) 2023 - 2024, Björn Darri Sigurðsson.
  */
+
 /**
  * Chinook demo.
  */
@@ -46,6 +47,6 @@ module is.codion.framework.demos.chinook {
 					with is.codion.framework.demos.chinook.domain.impl.ChinookImpl;
 	provides is.codion.common.rmi.server.Authenticator
 					with is.codion.framework.demos.chinook.server.ChinookAuthenticator;
-	provides is.codion.common.resource.Messages
-					with is.codion.framework.demos.chinook.model.ChinookMessages;
+	provides is.codion.common.resource.Resources
+					with is.codion.framework.demos.chinook.model.ChinookResources;
 }